requirements.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
certifi==2019.9.11
chardet==3.0.4
charset-normalizer==3.0.1
colorama==0.4.6
defusedxml==0.7.1
Django==3.2.18
django-filter==22.1
djangorestframework==3.14.0
exceptiongroup==1.1.0
flake8==5.0.4
httpie==3.2.1
idna==2.8
importlib-metadata==6.0.0
iniconfig==2.0.0
Markdown==3.4.1
markdown-it-py==2.1.0
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
packaging==23.0
pluggy==1.0.0
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.14.0
PySocks==1.7.1
pytest==7.2.1
pytz==2022.7.1
requests==2.22.0
requests-toolbelt==0.10.1
rich==13.3.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.25.6
zipp==3.13.0
